fix: add id verification in repository

+++ src/controllers/studentController.ts
@@ -14,7 +14,10 @@ export class StudentController {
     }
     getStudentById = async (req: Request, res: Response): Promise<Response> => {
         try{
-            const id = parseInt(req.params.id) as number;
+            const id = req.params.id ? parseInt(req.params.id): null;
+            if (id === null || isNaN(id)){
+                return res.status(400).json({message: 'Invalid ID'});
+            }
 
             const student = await studentRepository.findById(id.toString());
             if(!student){
@@ -40,7 +43,10 @@ export class StudentController {
     }
     deleteStudentById= async (req: Request, res: Response): Promise<Response> => {
         try{
-            const id= parseInt(req.params.id) as number;
+            const id= req.params.id ? parseInt(req.params.id): null;
+            if (id === null || isNaN(id)){
+                return res.status(400).json({message: 'Invalid ID'});
+            }
             const success = await studentRepository.delete(id.toString());
             if (!success){
                 return res.status(404).json({message: 'Student not found'});
@@ -52,7 +58,10 @@ export class StudentController {
     }
     updateStudentName= async (req: Request, res: Response): Promise<Response> => {
         try{
-            const id = parseInt(req.params.id) as number;
+            const id = req.params.id ? parseInt(req.params.id): null;
+            if (id === null || isNaN(id)){
+                return res.status(400).json({message: 'Invalid ID'});
+            }
             const {name} = req.body;
 
             if (!name || !name.length) return res.status(400).json({message: 'Invalid body response'});
@@ -96,7 +105,10 @@ export class StudentController {
     }
     updateStudentPhone= async (req: Request, res: Response): Promise<Response> => {
         try{
-            const id= parseInt(req.params.id) as number;
+            const id= req.params.id ? parseInt(req.params.id): null;
+            if (id === null || isNaN(id)){
+                return res.status(400).json({message: 'Invalid ID'});
+            }
             const {phone_number} = req.body;
 
             if (!phone_number || !phone_number.toString().length) return res.status(400).json({message: 'Invalid body response'});
@@ -109,3 +121,4 @@ export class StudentController {
         }
     }
 };
+
